move multisig example to /scripts/multisig
