+++ examples/deployment/assets/gold-john-bob.mlsig
@@ -16,15 +16,15 @@ LogicSig: {
   "msig": {
     "subsig": [
       {
-        "pk": "STICGNGDJKFGSSHHXALD6VBEJ5XLWL7QK5UIGV3DE2ONEXNFB6ZSOMGP6Y",
-        "s": "9xBRPfMoE/RFjHT1KP+aWBrEIGfTLRwJWT+G18eMcj7/YXrrgb3xhQeIn4ubh5UfVVXCjadv6j8vGwLFl02yCg=="
+        "pk": "M7VR2MGHI35EG2NMYOF3X337636PIOFVSP2HNIFUKAG7WW6BDWDCA3E2DA",
+        "s": "vEbfm3wp9ehSu45eMAyMj2fmMDTL2OPsgUS2ofcI53zgInTmEN2dpftt5mrFtLAUgyGTSMxLvAcgChqzSbWSDA=="
       },
       {
-        "pk": "U6NEV736OHP7V7QHXKZFP6QAPE5PMYJOYIPWT2VZCWIMWGJS62ZK7XLKCE",
-        "s": "cnKNV7ZvKUMwOvLPNWxDlpxEi9pgQrFLcUXlecI2w1kT+kb2K/WR20v3GF8MDhEDLKO//O98l8cUGQvsLw8RBg=="
+        "pk": "2UBZKFR6RCZL7R24ZG327VKPTPJUPFM6WTG7PJG2ZJLU234F5RGXFLTAKA",
+        "s": "dtxIk/TpH1x5wQqZheEKCRINnbhwX/5/nBku6TShJh4ccfhXNEpHa136cYxAdafWKooa1j5ZVoxL/a2qSYd/Cw==" 
       },
       {
-        "pk": "Z2SFZ7X2RAKTU4JYGUYYE7LDIE6M7NYCMUQEAIKII4UDT5KYE4IJNZAGCU"
+        "pk": "2ILRL5YU3FZ4JDQZQVXEZUYKEWF7IEIGRRCPCMI36VKSGDMAS6FHSBXZDQ"
       }
     ],
     "thr": 2,
